refactor(consensus/adaptor): split router.go god file, dedup tx-set fill (#897)
Phase B structural split of the 2,790-line router.go along its seams
(pure code motion — no behavior change, verified by `go test -race`):

- router_dispatch.go — message dispatch + proposal/validation/manifest/
  transaction ingest
- router_serve.go    — ledger / tx-set serving (handleGetLedger, serveTxSet,
  SHAMap reply building)
- router_txset.go    — the tx-set acquisition state machine
- router_catchup.go  — status-change handling, ledger acquisition,
  replay-delta, inbound-ledger completion

router.go now holds only the Router type, construction, setters, Run, and
the maintenance tick.

Also extracts fillTxSetFromLocalPool, the previously-duplicated local-pool
fill loop shared by the inbound and timer acquisition paths. The two paths'
throttle/attempt-cap control flow is intentionally left distinct: they
differ in per-peer progress accounting and merging them is a behavioral
change that needs multi-node soak validation, out of scope here.

Refs #897
