+++ internal/consensus/adaptor/router.go
@@ -5,22 +5,14 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"slices"
 	"sync"
 	"time"
 
 	"github.com/LeJamon/go-xrpl/internal/consensus"
-	"github.com/LeJamon/go-xrpl/internal/ledger"
-	"github.com/LeJamon/go-xrpl/internal/ledger/header"
 	"github.com/LeJamon/go-xrpl/internal/ledger/inbound"
-	"github.com/LeJamon/go-xrpl/internal/ledger/openledger"
-	"github.com/LeJamon/go-xrpl/internal/ledger/service"
 	"github.com/LeJamon/go-xrpl/internal/manifest"
 	"github.com/LeJamon/go-xrpl/internal/peermanagement"
-	"github.com/LeJamon/go-xrpl/internal/peermanagement/message"
 	validatorlist "github.com/LeJamon/go-xrpl/internal/validator/list"
-	"github.com/LeJamon/go-xrpl/protocol"
-	"github.com/LeJamon/go-xrpl/shamap"
 )
 
 // inboundReplayDeltaTickInterval drives the periodic check for
@@ -148,63 +140,6 @@ type Router struct {
 	// guarantee implicitly because online-delete physically removed the data.
 	// Nil when online-delete is off, leaving acquisition/serving unrestricted.
 	floor MinimumOnlineFloor
-}
-
-type txSetAcquireState struct {
-	txMap      *shamap.SHAMap
-	startedAt  time.Time
-	lastUpdate time.Time
-
-	// Retry bookkeeping. lastRequest is when we most recently broadcast a
-	// RequestTxSetMissingNodes; attempts counts those broadcasts so we can
-	// surface failure after a cap rather than spinning forever.
-	// peerNonProgress tracks consecutive TMLedgerData responses from a peer
-	// that failed to extend the SHAMap; peers at or over the per-peer
-	// threshold are skipped during the next broadcast.
-	lastRequest     time.Time
-	attempts        int
-	peerNonProgress map[uint64]int
-}
-
-// 60s covers a consensus round (~15s) plus retries with margin while
-// bounding memory under a stalled consumer.
-const txSetAcquireTTL = 60 * time.Second
-
-// txSetRetryKnobs collects the tunable parameters of the tx-set acquire
-// retry loop, which is event-driven (one tick per inbound TMLedgerData).
-//
-//   - MinInterval: minimum spacing between successive broadcasts for
-//     the same acquisition (250ms), so a chatty peer can't drive the
-//     cadence faster.
-//   - MaxAttempts: hard cap on broadcasts per acquisition (20) before
-//     the acquire gives up.
-//   - PeerNonProgressThreshold: consecutive non-progressing
-//     TMLedgerData replies from one peer before it is skipped on the
-//     next broadcast. 3 is small enough to react quickly to a truly
-//     stuck peer and large enough to ride out a transient empty reply.
-type txSetRetryKnobs struct {
-	MinInterval              time.Duration
-	MaxAttempts              int
-	PeerNonProgressThreshold int
-}
-
-func defaultTxSetRetryKnobs() txSetRetryKnobs {
-	return txSetRetryKnobs{
-		MinInterval:              250 * time.Millisecond,
-		MaxAttempts:              20,
-		PeerNonProgressThreshold: 3,
-	}
-}
-
-// SetTxSetRetryKnobsForTest overrides the tx-set retry knobs on this
-// Router. Tests use it to dial timings down so they don't sleep for the
-// production throttle window. The lock matches the read in handleTxSetData
-// so racing this against an active inbox goroutine is safe under -race;
-// production is not expected to call this.
-func (r *Router) SetTxSetRetryKnobsForTest(knobs txSetRetryKnobs) {
-	r.txSetAcquireMu.Lock()
-	defer r.txSetAcquireMu.Unlock()
-	r.txSetRetryKnobs = knobs
 }
 
 // messageDedupTTL is how long a proposal/validation hash is
@@ -465,108 +400,6 @@ func (r *Router) maintenanceTick() {
 	r.retryStalledTxSetAcquires()
 }
 
-func (r *Router) handleMessage(msg *peermanagement.InboundMessage) {
-	msgType := message.MessageType(msg.Type)
-
-	switch msgType {
-	case message.TypeProposeLedger:
-		r.handleProposal(msg)
-	case message.TypeValidation:
-		r.handleValidation(msg)
-	case message.TypeTransaction:
-		r.handleTransaction(msg)
-	case message.TypeHaveSet:
-		r.handleHaveSet(msg)
-	case message.TypeStatusChange:
-		r.handleStatusChange(msg)
-	case message.TypeGetLedger:
-		r.handleGetLedger(msg)
-	case message.TypeLedgerData:
-		r.handleLedgerData(msg)
-	case message.TypeGetObjects:
-		// Only fetch-pack REPLIES reach the router; the overlay serves
-		// otFETCH_PACK requests inline and forwards replies here (see
-		// handleGetObjectsMessage). handleFetchPackReply ignores anything
-		// that isn't an otFETCH_PACK reply.
-		r.handleFetchPackReply(msg)
-	case message.TypeReplayDeltaResponse:
-		r.handleReplayDeltaResponse(msg)
-	case message.TypeProofPathResponse:
-		r.handleProofPathResponse(msg)
-	case message.TypeManifests:
-		r.handleManifests(msg)
-	case message.TypeValidatorList:
-		r.handleValidatorList(msg)
-	case message.TypeValidatorListCollection:
-		r.handleValidatorListCollection(msg)
-	default:
-	}
-}
-
-// handleManifests ingests a TMManifests frame. For each serialized
-// manifest in the list: deserialize, apply to the cache, and — on
-// Accepted — relay the single-manifest frame to every peer except the
-// origin.
-//
-// Decode failures attribute "manifest-decode" badData to the sender. A
-// mix of valid and invalid entries in the same frame results in the
-// valid ones being applied; the frame isn't rejected wholesale.
-func (r *Router) handleManifests(msg *peermanagement.InboundMessage) {
-	if r.manifests == nil {
-		// Cache not wired (tests or minimal configs) — silently drop.
-		return
-	}
-
-	decoded, err := message.Decode(message.TypeManifests, msg.Payload)
-	if err != nil {
-		r.logger.Warn("failed to decode manifests frame", "error", err, "peer", msg.PeerID)
-		r.adaptor.IncPeerBadData(uint64(msg.PeerID), "manifests-decode")
-		return
-	}
-	mfs, ok := decoded.(*message.Manifests)
-	if !ok || len(mfs.List) == 0 {
-		return
-	}
-
-	for _, wire := range mfs.List {
-		parsed, err := manifest.Deserialize(wire.STObject)
-		if err != nil {
-			r.logger.Debug("manifest parse failed",
-				"error", err, "peer", msg.PeerID)
-			r.adaptor.IncPeerBadData(uint64(msg.PeerID), "manifest-parse")
-			continue
-		}
-		switch d := r.manifests.ApplyManifest(parsed); d {
-		case manifest.Accepted:
-			r.relayManifest(msg.PeerID, wire.STObject)
-		case manifest.Invalid, manifest.BadMasterKey, manifest.BadEphemeralKey:
-			// Charge the sender — they gave us a manifest that
-			// passed structural parse but failed the cache's
-			// invariants (signature, key reuse, etc.).
-			r.adaptor.IncPeerBadData(uint64(msg.PeerID), "manifest-"+d.String())
-		case manifest.Stale:
-			// Expected and harmless: a peer gossiped a manifest we
-			// already have at equal or higher seq. No action.
-		}
-	}
-}
-
-// relayManifest rebroadcasts a single accepted manifest to every peer
-// except the origin. Wraps the serialized STObject in a TMManifests
-// frame (a list of one). Shares its framing with the local-manifest
-// emission paths in manifest_emit.go.
-func (r *Router) relayManifest(exceptPeer peermanagement.PeerID, serialized []byte) {
-	if r.overlay == nil {
-		return
-	}
-	frame, err := encodeManifestsFrame(serialized)
-	if err != nil {
-		r.logger.Warn("failed to encode manifest relay frame", "error", err)
-		return
-	}
-	_ = r.overlay.BroadcastExcept(exceptPeer, frame)
-}
-
 // Bounds used to reject malformed TMProposeSet / TMValidation frames
 // before they reach the engine. Out-of-range values get feeInvalidData
 // attributed to the sender.
@@ -578,515 +411,6 @@ const (
 	signatureMinLen = 64
 	signatureMaxLen = 72
 )
-
-func (r *Router) handleProposal(msg *peermanagement.InboundMessage) {
-	decoded, err := message.Decode(message.TypeProposeLedger, msg.Payload)
-	if err != nil {
-		r.logger.Warn("failed to decode proposal", "error", err, "peer", msg.PeerID)
-		r.adaptor.IncPeerBadData(uint64(msg.PeerID), "proposal-decode")
-		return
-	}
-	proposeSet, ok := decoded.(*message.ProposeSet)
-	if !ok {
-		return
-	}
-
-	// Bounds checks BEFORE the engine sees the frame, so a peer can't
-	// cost-free spam oversized or implausibly-hoppy consensus traffic.
-	if badField, ok := validateProposeBounds(proposeSet); !ok {
-		r.logger.Debug("dropping malformed proposal",
-			"peer", msg.PeerID, "bad_field", badField)
-		r.adaptor.IncPeerBadData(uint64(msg.PeerID), "proposal-malformed-"+badField)
-		return
-	}
-
-	proposal := ProposalFromMessage(proposeSet)
-	r.resolveMasterNodeID(&proposal.NodeID, proposal.SigningPubKey)
-	originPeer := uint64(msg.PeerID)
-
-	// Record duplicate-status + last-sighting BEFORE OnProposal.
-	// Hash the DECODED fields via hashProposalSuppression. Hashing
-	// the raw protobuf envelope would desync dedup from peers
-	// that see the same message with different optional-field framing
-	// (e.g., deprecated `hops` included or omitted) — same semantic
-	// proposal, but different byte payload.
-	//
-	// Stash the hash on the Proposal so the downstream relay path
-	// can thread it to Overlay's reverse index without recomputing.
-	suppressionHash := hashProposalSuppression(proposal)
-	proposal.SuppressionHash = suppressionHash
-	firstSeen, lastSeen := r.messageSeen.observe(suppressionHash)
-
-	// Drop duplicates before the engine path (re-running OnProposal
-	// just re-verifies ECDSA). Still feed the IDLED-gated relay slot
-	// on dupes for squelch accounting.
-	//
-	// Deliberate deviation: rippled tracks suppression per (hash, peer),
-	// re-running the handler so per-peer slot entries grow on each new
-	// sender. Our dedup is hash-only, so a second peer's copy is dropped
-	// at the gate. Quorum/position tracking unaffected (first arrival
-	// counts the validator); reduce-relay accuracy is partly compensated
-	// via PeersThatHave + UpdateRelaySlot below.
-	if !firstSeen {
-		if time.Since(lastSeen) < peermanagement.Idled {
-			seenPeers := r.adaptor.PeersThatHave(suppressionHash)
-			r.adaptor.UpdateRelaySlot(proposal.SigningPubKey[:], originPeer, seenPeers)
-		}
-		return
-	}
-
-	if err := r.engine.OnProposal(proposal, originPeer); err != nil {
-		r.logger.Debug("engine rejected proposal", "error", err, "peer", msg.PeerID)
-		return
-	}
-}
-
-func (r *Router) handleValidation(msg *peermanagement.InboundMessage) {
-	decoded, err := message.Decode(message.TypeValidation, msg.Payload)
-	if err != nil {
-		r.logger.Warn("failed to decode validation", "error", err, "peer", msg.PeerID)
-		r.adaptor.IncPeerBadData(uint64(msg.PeerID), "validation-decode")
-		return
-	}
-	val, ok := decoded.(*message.Validation)
-	if !ok {
-		return
-	}
-
-	validation, err := ValidationFromMessage(val)
-	if err != nil {
-		r.logger.Warn("failed to parse validation", "error", err, "peer", msg.PeerID)
-		r.adaptor.IncPeerBadData(uint64(msg.PeerID), "validation-parse")
-		return
-	}
-	r.resolveMasterNodeID(&validation.NodeID, validation.SigningPubKey)
-
-	// Post-parse bounds: the validation struct must carry sane hash
-	// and signature sizes. Same rationale as in handleProposal.
-	if badField, ok := validateValidationBounds(validation); !ok {
-		r.logger.Debug("dropping malformed validation",
-			"peer", msg.PeerID, "bad_field", badField)
-		r.adaptor.IncPeerBadData(uint64(msg.PeerID), "validation-malformed-"+badField)
-		return
-	}
-
-	originPeer := uint64(msg.PeerID)
-
-	// Observe-before-engine for consistent duplicate accounting. Hash
-	// the INNER STValidation blob carried in TMValidation.validation.
-	// Hashing the TMValidation envelope instead would desync dedup the
-	// same way handleProposal would if it hashed the TMProposeSet
-	// envelope: deprecated outer fields vary, inner canonical blob does
-	// not. We use the raw inbound bytes here — NOT a re-serialized copy
-	// — so a lossy or reordered round-trip can't silently diverge the
-	// hash. Stash the hash on the Validation so the downstream relay
-	// path can thread it to Overlay's reverse index without recomputing.
-	suppressionHash := hashValidationSuppression(val.Validation)
-	validation.SuppressionHash = suppressionHash
-	firstSeen, lastSeen := r.messageSeen.observe(suppressionHash)
-
-	// Drop duplicates before the engine path (re-running OnValidation
-	// just re-verifies ECDSA, dominating CPU under gossip fan-out).
-	// Still update the relay slot for squelch accounting.
-	//
-	// Deliberate deviation: rippled's per-(hash, peer) suppression
-	// re-processes new senders; our hash-only dedup drops them at the
-	// gate. See handleProposal for the full rationale.
-	if !firstSeen {
-		if time.Since(lastSeen) < peermanagement.Idled {
-			seenPeers := r.adaptor.PeersThatHave(suppressionHash)
-			r.adaptor.UpdateRelaySlot(validation.SigningPubKey[:], originPeer, seenPeers)
-		}
-		return
-	}
-
-	if err := r.engine.OnValidation(validation, originPeer); err != nil {
-		// A same-seq double-sign (conflicting/multiple) is attributable
-		// bad data — charge the delivering peer. Mirrors rippled surfacing
-		// ValStatus::conflicting/multiple from Validations::add.
-		var bv *consensus.ByzantineValidationError
-		if errors.As(err, &bv) {
-			r.adaptor.IncPeerBadData(originPeer, "validation-"+bv.Reason)
-		}
-		r.logger.Info("engine rejected validation",
-			"t", "consensus",
-			"event", "validation-rejected",
-			"error", err.Error(),
-			"peer", msg.PeerID)
-		return
-	}
-	r.logger.Info("inbound validation accepted",
-		"t", "consensus",
-		"event", "validation-recv",
-		"peer", msg.PeerID,
-		"seq", validation.LedgerSeq,
-		"hash_short", fmt.Sprintf("%x", validation.LedgerID[:8]))
-
-	// Per-validation catch-up acquire on EVERY trusted current
-	// validation, not only at quorum. Under sustained load a node that
-	// falls one ledger behind enters the wrongLedger chase loop holding
-	// no position (our_pos_seq=0); with only 3 of the 4-quorum trusted
-	// validators on the network tip, the quorum-gated stash acquire
-	// (armValidationStashAcquisition) never fires, so the node never
-	// fetches the tip the network is converging on and the chain stalls
-	// below quorum until a slow periodic sweep recovers it. Acquiring on
-	// each trusted validation breaks that loop.
-	r.maybeAcquireFromValidation(validation, originPeer)
-}
-
-// resolveMasterNodeID looks the inbound signing pubkey up in the
-// manifest cache and, when a manifest binds it to a master pubkey,
-// rewrites *nid to CalcNodeID(masterKey). In the absence of a manifest
-// mapping the parser's initial CalcNodeID(signingKey) value is
-// preserved untouched, so non-rotated validators still round-trip
-// through the engine on the signing-derived NodeID.
-//
-// The manifest cache is installed on the router via SetManifestCache
-// before Run(). When the cache is nil (tests constructing a bare
-// router), this is a no-op and the parser default stands.
-func (r *Router) resolveMasterNodeID(nid *consensus.NodeID, signing consensus.SigningPubKey) {
-	if r.manifests == nil {
-		return
-	}
-	master := r.manifests.GetMasterKey([33]byte(signing))
-	// GetMasterKey returns the input unchanged when no manifest has
-	// bound this signing key to a master — leave nid alone in that
-	// case so we don't redundantly rehash.
-	if master == [33]byte(signing) {
-		return
-	}
-	*nid = consensus.CalcNodeID(master)
-}
-
-// validateProposeBounds returns ("", true) when the decoded ProposeSet
-// is within bounds; returns (field_label, false) on the first violation
-// so the caller can attribute the charge with a specific reason.
-func validateProposeBounds(p *message.ProposeSet) (string, bool) {
-	if p == nil {
-		return "nil", false
-	}
-	if len(p.PreviousLedger) != 32 {
-		return "prev-ledger-size", false
-	}
-	if len(p.CurrentTxHash) != 32 {
-		return "txset-size", false
-	}
-	if n := len(p.Signature); n < signatureMinLen || n > signatureMaxLen {
-		return "sig-size", false
-	}
-	// Proposal pubkeys must be compressed secp256k1 (0x02/0x03 prefix).
-	// ed25519 validators (0xED prefix) are not allowed in propose-set.
-	// The length-only check would pass a 33-byte ed25519 key (0xED || 32
-	// bytes), letting the peer slip through without attribution, so the
-	// prefix gate runs alongside the size gate.
-	if len(p.NodePubKey) != 33 {
-		return "pubkey-size", false
-	}
-	if p.NodePubKey[0] != 0x02 && p.NodePubKey[0] != 0x03 {
-		return "pubkey-type", false
-	}
-	return "", true
-}
-
-// validateValidationBounds returns ("", true) when the parsed
-// Validation has sane lengths on the post-decode struct fields. Same
-// attribution contract as validateProposeBounds.
-func validateValidationBounds(v *consensus.Validation) (string, bool) {
-	if v == nil {
-		return "nil", false
-	}
-	if v.LedgerID == (consensus.LedgerID{}) {
-		return "ledger-hash-zero", false
-	}
-	if v.SigningPubKey == (consensus.SigningPubKey{}) {
-		return "signing-pubkey-zero", false
-	}
-	if n := len(v.Signature); n < signatureMinLen || n > signatureMaxLen {
-		return "sig-size", false
-	}
-	return "", true
-}
-
-func (r *Router) handleTransaction(msg *peermanagement.InboundMessage) {
-	decoded, err := message.Decode(message.TypeTransaction, msg.Payload)
-	if err != nil {
-		r.logger.Warn("failed to decode transaction", "error", err, "peer", msg.PeerID)
-		return
-	}
-	txMsg, ok := decoded.(*message.Transaction)
-	if !ok {
-		r.logger.Warn("decoded transaction has unexpected type",
-			"peer", msg.PeerID,
-			"got", fmt.Sprintf("%T", decoded))
-		return
-	}
-
-	blob := TransactionFromMessage(txMsg)
-	if len(blob) == 0 {
-		r.logger.Warn("inbound transaction has empty blob",
-			"peer", msg.PeerID,
-			"status", txMsg.Status)
-		return
-	}
-
-	// Peer-relay path — the originating peer manages its own resends,
-	// so we don't pin the blob in our LocalTxs held pool.
-	res, err := r.adaptor.SubmitPendingTx(blob, false)
-	r.logger.Info("inbound tx accepted into pending pool",
-		"t", "consensus",
-		"event", "tx-inbound",
-		"peer", msg.PeerID,
-		"blob_size", len(blob),
-		"status", txMsg.Status,
-	)
-
-	// Relay immediately on the inbound job, not one ledger later via
-	// OpenLedger.Accept's once-per-LCL callback; that one-ledger lag is a
-	// direct contributor to tx-propagation latency.
-	//
-	// Gate: relay only for the applied-or-terQUEUED case. openledger.Submit
-	// folds both terQUEUED and tec into ResultSuccess, so ResultSuccess is
-	// the exact superset that should relay. ResultRetry (non-queued ter*)
-	// and ResultFailure (tef/tem/tel) do NOT relay.
-	if err == nil && res == openledger.ResultSuccess {
-		r.relayTransaction(msg.PeerID, blob)
-	}
-}
-
-// relayTransaction rebroadcasts an accepted peer-originated TMTransaction,
-// excluding the originating peer.
-//
-// The outbound wire shape: status normalized to tsCURRENT (the inbound
-// peer's claimed status is informational only) and receivetimestamp
-// freshly stamped from the local Ripple clock.
-//
-// Overlay.RelayTransaction applies reduce-relay peer selection: the full
-// frame goes to a subset of peers and the rest learn of the tx via the
-// TMHaveTransactions announce. We don't consult a separate suppression
-// set for the multi-hop case because de-dup happens implicitly via
-// openledger.Submit's view.TxExists pre-filter: a duplicate arrival from
-// another peer classifies as ResultFailure and the relay gate above never
-// fires. Excluding the origin is the minimum correctness boundary —
-// without it the originator would receive its own packet back and either
-// re-charge us bandwidth or, in a 2-peer cycle, oscillate indefinitely.
-func (r *Router) relayTransaction(except peermanagement.PeerID, blob []byte) {
-	if r.overlay == nil {
-		return
-	}
-	out := &message.Transaction{
-		RawTransaction:   blob,
-		Status:           message.TxStatusCurrent,
-		ReceiveTimestamp: uint64(time.Now().Unix() - protocol.RippleEpochUnix),
-	}
-	frame, err := encodeFrame(message.TypeTransaction, out)
-	if err != nil {
-		r.logger.Warn("relay transaction frame build failed", "error", err)
-		return
-	}
-	// Reduce-relay peer selection: relays the full frame to a subset of
-	// peers and lets the rest learn via the TMHaveTransactions announce.
-	r.overlay.RelayTransaction(except, frame)
-}
-
-func (r *Router) handleHaveSet(msg *peermanagement.InboundMessage) {
-	decoded, err := message.Decode(message.TypeHaveSet, msg.Payload)
-	if err != nil {
-		r.logger.Warn("failed to decode have_set", "error", err, "peer", msg.PeerID)
-		return
-	}
-	hts, ok := decoded.(*message.HaveTransactionSet)
-	if !ok {
-		return
-	}
-
-	txSetID, status := HaveSetFromMessage(hts)
-
-	switch status {
-	case message.TxSetStatusHave:
-		// Peer has a tx set we might need — if the engine is waiting for it,
-		// we could request the full set. For now, just log.
-		r.logger.Debug("peer has txset", "txset", txSetID, "peer", msg.PeerID)
-	case message.TxSetStatusNeed:
-		// Peer needs a tx set we might have — check cache and respond.
-		if ts, ok := r.adaptor.txSetCache.Get(txSetID); ok {
-			// We have it — notify the engine with the tx set data
-			if err := r.engine.OnTxSet(ts.ID(), ts.Txs()); err != nil {
-				r.logger.Debug("engine rejected txset", "error", err)
-			}
-		}
-	}
-}
-
-func (r *Router) handleGetLedger(msg *peermanagement.InboundMessage) {
-	decoded, err := message.Decode(message.TypeGetLedger, msg.Payload)
-	if err != nil {
-		r.logger.Warn("failed to decode get_ledger", "error", err, "peer", msg.PeerID)
-		return
-	}
-	req, ok := decoded.(*message.GetLedger)
-	if !ok {
-		return
-	}
-
-	r.logger.Debug("peer requests ledger",
-		"peer", msg.PeerID,
-		"itype", req.InfoType,
-		"seq", req.LedgerSeq,
-		"hash_len", len(req.LedgerHash),
-	)
-
-	// For a liTS_CANDIDATE request, ledger_hash carries the tx-set ID and
-	// the response is TMLedgerData{type=liTS_CANDIDATE, ...}.
-	if req.InfoType == message.LedgerInfoTsCandidate {
-		r.serveTxSet(msg.PeerID, req)
-		return
-	}
-
-	// liBASE / liAS_NODE / liTX_NODE all operate on a stored ledger; only
-	// these three plus liTS_CANDIDATE are served, so anything else is
-	// dropped.
-	switch req.InfoType {
-	case message.LedgerInfoBase, message.LedgerInfoAsNode, message.LedgerInfoTxNode:
-	default:
-		return
-	}
-
-	svc := r.adaptor.LedgerService()
-	if svc == nil {
-		return
-	}
-
-	// Load-shed ledger-BODY requests under load. The liTS_CANDIDATE branch
-	// above is intentionally exempt, because consensus liveness depends on
-	// tx-set acquisition always being served — so this gate runs only for
-	// liBASE / liAS_NODE / liTX_NODE.
-	loadedLocal := false
-	if ft := svc.FeeTrack(); ft != nil {
-		loadedLocal = ft.IsLoadedLocal()
-	}
-	if r.adaptor.ShouldShedLedgerRequest(uint64(msg.PeerID), loadedLocal) {
-		r.logger.Debug("get_ledger shed under load",
-			"peer", msg.PeerID, "itype", req.InfoType)
-		return
-	}
-
-	var l *ledger.Ledger
-	if len(req.LedgerHash) == 32 {
-		var hash [32]byte
-		copy(hash[:], req.LedgerHash)
-		l, err = svc.GetLedgerByHash(hash)
-	} else if req.LedgerSeq > 0 {
-		l, err = svc.GetLedgerBySequence(req.LedgerSeq)
-	} else {
-		l = svc.GetClosedLedger()
-	}
-	if err != nil || l == nil {
-		return
-	}
-
-	// Refuse to serve a ledger online-delete has reclaimed: rippled can't
-	// serve what its store deleted, so neither do we (the in-memory history
-	// window may still hold it). The liTS_CANDIDATE tx-set path returned
-	// above is exempt — consensus liveness depends on it always serving.
-	if r.belowFloor(l.Sequence()) {
-		r.logger.Debug("get_ledger declined: below online-delete floor",
-			"peer", msg.PeerID, "seq", l.Sequence(), "floor", r.floor.MinimumOnline())
-		return
-	}
-
-	hash := l.Hash()
-	var nodes []message.LedgerNode
-	switch req.InfoType {
-	case message.LedgerInfoBase:
-		nodes = r.buildLedgerBaseNodes(l)
-	case message.LedgerInfoAsNode:
-		nodes = r.serveLedgerMapNodes(l.StateMapSnapshot, req, msg.PeerID, "state")
-	case message.LedgerInfoTxNode:
-		nodes = r.serveLedgerMapNodes(l.TxMapSnapshot, req, msg.PeerID, "tx")
-	}
-	if len(nodes) == 0 {
-		return
-	}
-
-	resp := &message.LedgerData{
-		LedgerHash:    hash[:],
-		LedgerSeq:     l.Sequence(),
-		InfoType:      req.InfoType,
-		Nodes:         nodes,
-		RequestCookie: uint32(req.RequestCookie),
-	}
-
-	frame, err := encodeFrame(message.TypeLedgerData, resp)
-	if err != nil {
-		r.logger.Warn("failed to encode ledger_data response", "error", err)
-		return
-	}
-
-	if err := r.adaptor.SendToPeer(uint64(msg.PeerID), frame); err != nil {
-		r.logger.Debug("failed to send ledger_data to peer", "error", err, "peer", msg.PeerID)
-	}
-}
-
-// buildLedgerBaseNodes builds the liBASE reply node set: node[0] is the
-// wire header (no trailing hash); node[1] is the account-state root when
-// the state tree is non-empty; node[2] is the transaction root when the
-// ledger has transactions (txHash and tx-map hash both non-zero).
-func (r *Router) buildLedgerBaseNodes(l *ledger.Ledger) []message.LedgerNode {
-	nodes := []message.LedgerNode{{NodeData: header.AddRaw(l.Header(), false)}}
-
-	stateHash, err := l.StateMapHash()
-	if err != nil || stateHash == ([32]byte{}) {
-		return nodes
-	}
-	stateMap, err := l.StateMapSnapshot()
-	if err != nil {
-		r.logger.Warn("ledger base: state snapshot failed", "error", err)
-		return nodes
-	}
-	stateRoot, err := stateMap.SerializeRoot()
-	if err != nil {
-		r.logger.Warn("ledger base: state root serialize failed", "error", err)
-		return nodes
-	}
-	nodes = append(nodes, message.LedgerNode{NodeData: stateRoot})
-
-	if l.Header().TxHash == ([32]byte{}) {
-		return nodes
-	}
-	txHash, err := l.TxMapHash()
-	if err != nil || txHash == ([32]byte{}) {
-		return nodes
-	}
-	txMap, err := l.TxMapSnapshot()
-	if err != nil {
-		r.logger.Warn("ledger base: tx snapshot failed", "error", err)
-		return nodes
-	}
-	txRoot, err := txMap.SerializeRoot()
-	if err != nil {
-		r.logger.Warn("ledger base: tx root serialize failed", "error", err)
-		return nodes
-	}
-	return append(nodes, message.LedgerNode{NodeData: txRoot})
-}
-
-// serveLedgerMapNodes walks the requested SHAMap node IDs of a ledger's
-// state or transaction tree: fat nodes, QueryDepth levels deep, honouring
-// the soft/hard reply caps. fatLeaves is true here — only liTS_CANDIDATE
-// uses false.
-func (r *Router) serveLedgerMapNodes(snapshot func() (*shamap.SHAMap, error), req *message.GetLedger, peerID peermanagement.PeerID, label string) []message.LedgerNode {
-	m, err := snapshot()
-	if err != nil || m == nil {
-		r.logger.Debug("ledger node serve: map snapshot unavailable", "error", err, "peer", peerID)
-		return nil
-	}
-	queryDepth := int(req.QueryDepth)
-	if queryDepth == 0 {
-		queryDepth = defaultQueryDepth
-	}
-	return buildShaMapReplyNodes(m, req.NodeIDs, queryDepth, true, r.logger, peerID,
-		fmt.Sprintf("ledger %d %s", req.LedgerSeq, label))
-}
 
 // Ledger-data serve-path caps, shared across liTS_CANDIDATE, liAS_NODE,
 // and liTX_NODE replies. Soft cap stops starting new subtrees; hard cap
@@ -1105,1687 +429,7 @@ const shamapNodeIDLen = 33
 // level is harmless, too few stalls the requestor.
 const defaultQueryDepth = 2
 
-// serveTxSet replies to TMGetLedger{itype=liTS_CANDIDATE} with the tx set
-// encoded as TMLedgerData{type=liTS_CANDIDATE, ledger_hash=<txSetID>,
-// nodes=[<SHAMapNodeID, wire-serialized SHAMap node>...]}. For each
-// requested NodeID, walk QueryDepth levels via GetNodeFatByPath, honouring
-// soft/hard caps. Empty nodeids falls back to a full pre-order walk for
-// legacy goxrpl→goxrpl fixtures; real requestors always send at least the
-// root.
-func (r *Router) serveTxSet(peerID peermanagement.PeerID, req *message.GetLedger) {
-	if len(req.LedgerHash) != 32 {
-		return
-	}
-	var txSetID consensus.TxSetID
-	copy(txSetID[:], req.LedgerHash)
-
-	ts, ok := r.adaptor.txSetCache.Get(txSetID)
-	if !ok {
-		r.logger.Debug("peer requested tx-set we don't have",
-			"peer", peerID, "txset", fmt.Sprintf("%x", txSetID[:8]))
-		return
-	}
-
-	// SHAMap is non-nil for any TxSet that reached the cache —
-	// NewTxSet returns an error before stashing on shamap.New failure.
-	txMap := ts.shamap()
-
-	queryDepth := int(req.QueryDepth)
-	if queryDepth == 0 {
-		queryDepth = defaultQueryDepth
-	}
-	// fatLeaves is always false for liTS_CANDIDATE.
-	const fatLeaves = false
-
-	nodes := buildShaMapReplyNodes(txMap, req.NodeIDs, queryDepth, fatLeaves, r.logger, peerID,
-		fmt.Sprintf("txset %x", txSetID[:8]))
-
-	resp := &message.LedgerData{
-		LedgerHash:    req.LedgerHash,
-		LedgerSeq:     0, // tx-set responses carry no ledger seq
-		InfoType:      message.LedgerInfoTsCandidate,
-		Nodes:         nodes,
-		RequestCookie: uint32(req.RequestCookie),
-	}
-
-	frame, err := encodeFrame(message.TypeLedgerData, resp)
-	if err != nil {
-		r.logger.Warn("failed to encode tx-set response", "error", err)
-		return
-	}
-
-	if err := r.adaptor.SendToPeer(uint64(peerID), frame); err != nil {
-		r.logger.Debug("failed to send tx-set response", "error", err, "peer", peerID)
-		return
-	}
-	r.logger.Debug("served tx-set to peer",
-		"peer", peerID,
-		"txset", fmt.Sprintf("%x", txSetID[:8]),
-		"shamap_nodes", len(nodes),
-		"txs", ts.Size(),
-		"query_depth", queryDepth,
-		"requested_nodes", len(req.NodeIDs))
-}
-
-// buildShaMapReplyNodes builds the LedgerNode payload of a TMLedgerData reply
-// (liTS_CANDIDATE / liAS_NODE / liTX_NODE), honouring requested
-// NodeIDs/QueryDepth and soft/hard reply caps. label identifies the source map
-// for logging.
-func buildShaMapReplyNodes(
-	m *shamap.SHAMap,
-	requestedNodeIDs [][]byte,
-	queryDepth int,
-	fatLeaves bool,
-	logger logger,
-	peerID peermanagement.PeerID,
-	label string,
-) []message.LedgerNode {
-	if len(requestedNodeIDs) == 0 {
-		wireNodes, err := m.WalkWireNodes()
-		if err != nil {
-			logger.Warn("failed to walk SHAMap for serve",
-				"error", err, "peer", peerID, "map", label)
-			return nil
-		}
-		nodes := make([]message.LedgerNode, 0, len(wireNodes))
-		for _, n := range wireNodes {
-			if len(nodes) >= txSetHardMaxReplyNodes {
-				break
-			}
-			nodes = append(nodes, message.LedgerNode{NodeID: n.NodeID, NodeData: n.Data})
-		}
-		return nodes
-	}
-
-	nodes := make([]message.LedgerNode, 0)
-	for i, rawID := range requestedNodeIDs {
-		// Soft cap: stop starting new subtrees.
-		if len(nodes) >= txSetSoftMaxReplyNodes {
-			logger.Debug("shamap serve: soft-cap reached, stopping subtree iteration",
-				"peer", peerID, "map", label,
-				"nodes_so_far", len(nodes), "remaining_requested", len(requestedNodeIDs)-i)
-			break
-		}
-		path, depth, ok := parseSHAMapNodeID(rawID)
-		if !ok {
-			logger.Debug("shamap serve: bad SHAMapNodeID in request, skipping",
-				"peer", peerID, "map", label,
-				"node_idx", i, "len", len(rawID))
-			continue
-		}
-		subtree, err := m.GetNodeFatByPath(path, depth, queryDepth, fatLeaves)
-		if err != nil {
-			logger.Debug("shamap serve: GetNodeFatByPath failed, skipping",
-				"peer", peerID, "map", label,
-				"error", err.Error())
-			continue
-		}
-		for _, n := range subtree {
-			// Hard cap: truncate mid-subtree.
-			if len(nodes) >= txSetHardMaxReplyNodes {
-				logger.Debug("shamap serve: hard-cap reached, truncating subtree",
-					"peer", peerID, "map", label,
-					"nodes", len(nodes))
-				return nodes
-			}
-			nodes = append(nodes, message.LedgerNode{NodeID: n.NodeID, NodeData: n.Data})
-		}
-	}
-	return nodes
-}
-
-// parseSHAMapNodeID decodes the 33-byte wire representation into (path,
-// depth).
-func parseSHAMapNodeID(raw []byte) (path [32]byte, depth int, ok bool) {
-	if len(raw) != shamapNodeIDLen {
-		return path, 0, false
-	}
-	copy(path[:], raw[:32])
-	depth = int(raw[32])
-	if depth < 0 || depth > 64 {
-		return path, 0, false
-	}
-	return path, depth, true
-}
-
 type logger interface {
 	Debug(msg string, args ...any)
 	Warn(msg string, args ...any)
-}
-
-// learnTxFromLeaf submits the transaction carried by an acquired tx-set
-// leaf into the open-ledger pool and, on acceptance, actively relays it.
-// A tx-set leaf is a tnTRANSACTION_NM node whose wire form is
-// `tx_blob || WireTypeTransaction`; inner nodes and malformed data are
-// skipped by the trailing-type-byte check, and a tx the open ledger already
-// holds is not resubmitted. The submit is peer-sourced and the relay reuses
-// relayTransaction exactly as handleTransaction does for an inbound
-// TMTransaction (see handleTransaction), excluding originPeer as the tx's
-// source — so a set the node only holds transiently still pushes its novel
-// txs to peers instead of relying on the slower TMHaveTransactions announce.
-//
-// We deliberately do not keep a per-acquisition node cache: the acquired
-// node is already held in txMap and the missing-leaf local-fill (below)
-// re-sources tx leaves from the open-ledger pool, so a cache has no role
-// here.
-func (r *Router) learnTxFromLeaf(originPeer uint64, wire []byte) {
-	if len(wire) < 2 || wire[len(wire)-1] != protocol.WireTypeTransaction {
-		return
-	}
-	leaf, err := shamap.NewTransactionLeafFromWire(wire)
-	if err != nil {
-		return
-	}
-	item := leaf.Item()
-	if item == nil {
-		return
-	}
-	if r.adaptor.HasTx(consensus.TxID(item.Key())) {
-		return
-	}
-	if res, err := r.adaptor.SubmitPendingTx(item.Data(), false); err == nil && res == openledger.ResultSuccess {
-		r.relayTransaction(peermanagement.PeerID(originPeer), item.Data())
-	}
-}
-
-// txLeafWire frames a raw transaction blob as a SHAMap transaction-leaf
-// node: `tx_blob || WireTypeTransaction`. shamap.NewTransactionLeafFromWire
-// and the DeserializeNodeFromWire dispatch are the inverse.
-func txLeafWire(blob []byte) []byte {
-	wire := make([]byte, len(blob)+1)
-	copy(wire, blob)
-	wire[len(blob)] = byte(protocol.WireTypeTransaction)
-	return wire
-}
-
-// buildExcludedPeers returns the set of peer IDs whose consecutive
-// non-progress count has reached threshold, so the next missing-nodes
-// request can route around them. Returns nil when none qualify (a nil
-// map is a valid empty exclusion set for RequestTxSetMissingNodes).
-func buildExcludedPeers(peerNonProgress map[uint64]int, threshold int) map[uint64]bool {
-	var excluded map[uint64]bool
-	for pid, count := range peerNonProgress {
-		if count >= threshold {
-			if excluded == nil {
-				excluded = make(map[uint64]bool)
-			}
-			excluded[pid] = true
-		}
-	}
-	return excluded
-}
-
-// missingNodeIDs projects the wire NodeID bytes of every still-missing
-// SHAMap node, in order, for RequestTxSetMissingNodes.
-func missingNodeIDs(missing []shamap.MissingNode) [][]byte {
-	nodeIDs := make([][]byte, len(missing))
-	for i, m := range missing {
-		nodeIDs[i] = m.NodeID.Bytes()
-	}
-	return nodeIDs
-}
-
-// handleTxSetData consumes a TMLedgerData{type=liTS_CANDIDATE} response.
-// Each node is a SHAMap node (root/inner/leaf), not a raw transaction.
-// It accumulates nodes across responses, then either finishes (→
-// engine.OnTxSet) or requests missing nodes. State is keyed by tx-set ID
-// so partial responses can resume.
-//
-// originPeer is the peer ID of the sender, used to attribute non-progress
-// for per-peer de-prioritization.
-func (r *Router) handleTxSetData(ld *message.LedgerData, originPeer uint64) {
-	if len(ld.LedgerHash) != 32 {
-		return
-	}
-	var txSetID consensus.TxSetID
-	copy(txSetID[:], ld.LedgerHash)
-
-	r.txSetAcquireMu.Lock()
-	state, exists := r.txSetAcquire[txSetID]
-	if !exists {
-		txMap := shamap.New(shamap.TypeTransaction)
-		if err := txMap.StartSync(); err != nil {
-			r.txSetAcquireMu.Unlock()
-			r.logger.Info("tx-set sync: StartSync failed",
-				"t", "consensus", "event", "txset-reject",
-				"txset", fmt.Sprintf("%x", txSetID[:8]),
-				"error", err.Error())
-			return
-		}
-		state = &txSetAcquireState{
-			txMap:           txMap,
-			startedAt:       time.Now(),
-			peerNonProgress: make(map[uint64]int),
-		}
-		r.txSetAcquire[txSetID] = state
-	}
-	state.lastUpdate = time.Now()
-	r.sweepStaleTxSetAcquireLocked()
-	txMap := state.txMap
-	r.txSetAcquireMu.Unlock()
-
-	// Root NodeID is 33 zero bytes. AddRootNode is idempotent
-	// (ErrRootAlreadySet treated as success). rootAccepted feeds
-	// per-peer progress accounting so a peer whose reply contains
-	// only the root still counts as making progress — any successful
-	// add (root or non-root) is useful.
-	rootAccepted := false
-	for _, node := range ld.Nodes {
-		if !isShamapRootNodeID(node.NodeID) {
-			continue
-		}
-		err := txMap.AddRootNode([32]byte(txSetID), node.NodeData)
-		switch {
-		case err == nil, errors.Is(err, shamap.ErrRootAlreadySet):
-			rootAccepted = true
-		default:
-			r.logger.Info("tx-set sync: AddRootNode failed",
-				"t", "consensus", "event", "txset-reject",
-				"txset", fmt.Sprintf("%x", txSetID[:8]),
-				"error", err.Error())
-		}
-		break
-	}
-
-	// Use NodeID-based placement: path-based descent works when the
-	// peer's response contains nodes deeper than our currently-loaded
-	// layer of stubs. The previous hash-search approach
-	// (AddKnownNodeUnchecked) silently rejected every node it couldn't
-	// place beneath a loaded parent, producing the missing-nodes retry
-	// storm of issue #413.
-	//
-	// replyValid stays true unless any non-root node fails parse or
-	// AddKnownNodeByID: a single bad non-root poisons the WHOLE reply.
-	// We MUST reflect this all-or-nothing reply outcome in per-peer
-	// non-progress accounting so a peer trickling junk alongside one
-	// good node can't keep its counter pinned at zero.
-	added := 0
-	replyValid := true
-	for _, node := range ld.Nodes {
-		if isShamapRootNodeID(node.NodeID) {
-			continue
-		}
-		if len(node.NodeData) == 0 {
-			continue
-		}
-		parsedID, err := shamap.UnmarshalBinary(node.NodeID)
-		if err != nil {
-			replyValid = false
-			r.logger.Debug("tx-set sync: malformed node ID",
-				"t", "consensus", "event", "txset-node-reject",
-				"txset", fmt.Sprintf("%x", txSetID[:8]),
-				"node_id_len", len(node.NodeID),
-				"error", err.Error())
-			continue
-		}
-		if _, err := txMap.AddKnownNodeByID(parsedID, node.NodeData); err != nil {
-			replyValid = false
-			r.logger.Debug("tx-set sync: node rejected",
-				"t", "consensus", "event", "txset-node-reject",
-				"txset", fmt.Sprintf("%x", txSetID[:8]),
-				"node_id", fmt.Sprintf("%x", node.NodeID),
-				"node_data_len", len(node.NodeData),
-				"error", err.Error())
-			continue
-		}
-		added++
-		// Learn (submit + relay) the tx carried by this acquired leaf.
-		r.learnTxFromLeaf(originPeer, node.NodeData)
-	}
-
-	if err := txMap.FinishSync(); err != nil {
-		// Request the missing nodes. Before going to peers, fill any
-		// missing TX-leaf nodes from our own pending pool. For
-		// tnTRANSACTION_NM the leaf-node hash equals the tx ID, so a
-		// single tx-ID lookup resolves the leaf. This is a round-trip
-		// optimization, not a correctness requirement: a peer DOES return
-		// a leaf requested directly by node ID, but local sourcing avoids
-		// the extra request for a tx we already relayed.
-		missing := txMap.GetMissingNodes(256, nil)
-		if len(missing) == 0 {
-			r.deleteTxSetAcquire(txSetID)
-			r.logger.Info("tx-set sync: stuck",
-				"t", "consensus", "event", "txset-reject",
-				"txset", fmt.Sprintf("%x", txSetID[:8]),
-				"err", err.Error())
-			return
-		}
-
-		filledFromPool := 0
-		remaining := missing[:0]
-		for _, m := range missing {
-			blob, getErr := r.adaptor.GetTx(consensus.TxID(m.Hash))
-			if getErr != nil || len(blob) == 0 {
-				remaining = append(remaining, m)
-				continue
-			}
-			wire := txLeafWire(blob)
-			if addErr := txMap.AddKnownNode(m.Hash, wire); addErr != nil {
-				remaining = append(remaining, m)
-				continue
-			}
-			filledFromPool++
-		}
-
-		if filledFromPool > 0 {
-			if syncErr := txMap.FinishSync(); syncErr == nil {
-				// Tree complete after local fill — fall through to the
-				// leaf-walk + engine feed below.
-				r.logger.Info("tx-set sync: completed via local pool",
-					"t", "consensus", "event", "txset-local-fill",
-					"txset", fmt.Sprintf("%x", txSetID[:8]),
-					"filled", filledFromPool,
-					"non_root_added", added,
-				)
-			} else {
-				// Still incomplete — recompute remaining via the SHAMap
-				// since AddKnownNode may have revealed deeper holes.
-				remaining = txMap.GetMissingNodes(256, nil)
-			}
-		}
-
-		if len(remaining) > 0 {
-			// Throttle retries, cap total attempts, and route around
-			// peers that have repeatedly failed to extend the SHAMap.
-			// Without these guards a non-progressing peer triggers 100+
-			// retries/sec until the 60s TTL sweep fires. A reply counts
-			// as progress only when it was non-empty AND every non-root
-			// node parsed and added cleanly. Any single bad non-root →
-			// invalid reply → not progress for the originating peer.
-			madeProgress := replyValid && (added > 0 || rootAccepted)
-			r.txSetAcquireMu.Lock()
-			// Knobs are read under txSetAcquireMu so a concurrent
-			// SetTxSetRetryKnobsForTest (test-only API) can't tear a
-			// half-updated struct into the hot path.
-			knobs := r.txSetRetryKnobs
-			if originPeer != 0 {
-				if madeProgress {
-					state.peerNonProgress[originPeer] = 0
-				} else {
-					state.peerNonProgress[originPeer]++
-				}
-			}
-			if !state.lastRequest.IsZero() && time.Since(state.lastRequest) < knobs.MinInterval {
-				r.txSetAcquireMu.Unlock()
-				r.logger.Debug("tx-set sync: retry throttled",
-					"t", "consensus", "event", "txset-retry-throttle",
-					"txset", fmt.Sprintf("%x", txSetID[:8]),
-					"missing", len(remaining),
-				)
-				return
-			}
-			if state.attempts >= knobs.MaxAttempts {
-				attempts := state.attempts
-				delete(r.txSetAcquire, txSetID)
-				r.txSetAcquireMu.Unlock()
-				// Drop the entry rather than mark it permanently
-				// failed: if consensus is still proposing this set, the
-				// next inbound TMLedgerData should start a fresh acquire
-				// rather than be silently dropped for the TTL window.
-				r.logger.Info("tx-set sync: max attempts exceeded",
-					"t", "consensus", "event", "txset-reject",
-					"txset", fmt.Sprintf("%x", txSetID[:8]),
-					"attempts", attempts,
-					"missing", len(remaining),
-				)
-				return
-			}
-			state.attempts++
-			state.lastRequest = time.Now()
-			excluded := buildExcludedPeers(state.peerNonProgress, knobs.PeerNonProgressThreshold)
-			attempts := state.attempts
-			r.txSetAcquireMu.Unlock()
-
-			nodeIDs := missingNodeIDs(remaining)
-			r.logger.Info("tx-set sync: requesting missing nodes",
-				"t", "consensus", "event", "txset-retry",
-				"txset", fmt.Sprintf("%x", txSetID[:8]),
-				"non_root_added", added,
-				"filled_local", filledFromPool,
-				"missing", len(remaining),
-				"attempts", attempts,
-				"excluded_peers", len(excluded),
-			)
-			if reqErr := r.adaptor.RequestTxSetMissingNodes(txSetID, nodeIDs, excluded); reqErr != nil {
-				r.logger.Info("tx-set sync: missing-nodes request failed",
-					"t", "consensus", "event", "txset-reject",
-					"txset", fmt.Sprintf("%x", txSetID[:8]),
-					"error", reqErr.Error())
-			}
-			return
-		}
-		// fall through: tree complete via local fill
-	}
-
-	// Walk leaves into blobs, feed the engine, drop the acquire so dispute
-	// resolution flipping back to the same set starts fresh.
-	blobs := make([][]byte, 0, added+1)
-	if err := txMap.ForEach(func(item *shamap.Item) bool {
-		blobs = append(blobs, item.Data())
-		return true
-	}); err != nil {
-		r.deleteTxSetAcquire(txSetID)
-		return
-	}
-	r.deleteTxSetAcquire(txSetID)
-
-	r.logger.Info("received tx-set from peer",
-		"t", "consensus", "event", "txset-recv",
-		"txset", fmt.Sprintf("%x", txSetID[:8]),
-		"node_count", len(ld.Nodes),
-		"tx_count", len(blobs))
-
-	// Duplicate response after a completed acquire — no root, ForEach
-	// yields 0 items, engine would fail with "tx set ID mismatch". Drop.
-	if len(blobs) == 0 {
-		return
-	}
-
-	r.submitTxSetToEngine(txSetID, blobs)
-}
-
-func (r *Router) deleteTxSetAcquire(txSetID consensus.TxSetID) {
-	r.txSetAcquireMu.Lock()
-	delete(r.txSetAcquire, txSetID)
-	r.txSetAcquireMu.Unlock()
-}
-
-// submitTxSetToEngine feeds a completed tx-set's blobs to the engine. An
-// engine rejection is logged, not fatal — the engine re-checks the tx-set
-// ID, so a stale or duplicate set is dropped rather than corrupting state.
-func (r *Router) submitTxSetToEngine(txSetID consensus.TxSetID, blobs [][]byte) {
-	if err := r.engine.OnTxSet(txSetID, blobs); err != nil {
-		r.logger.Info("engine rejected tx-set",
-			"t", "consensus", "event", "txset-reject",
-			"error", err.Error(),
-			"txset", fmt.Sprintf("%x", txSetID[:8]),
-			"tx_count", len(blobs))
-	}
-}
-
-// MarkTxSetStillNeeded is the active re-arm hook fired every time
-// consensus re-asks for a tx-set via Adaptor.RequestTxSet. If an
-// in-flight acquisition for this set still exists, attempts and
-// lastRequest are cleared so the next inbound TMLedgerData broadcasts
-// immediately instead of being throttled or silently dropped past the
-// max-attempts cap. A no-op if the router has no entry for txSetID
-// (e.g. first request, or already completed and swept).
-func (r *Router) MarkTxSetStillNeeded(txSetID consensus.TxSetID) {
-	r.txSetAcquireMu.Lock()
-	defer r.txSetAcquireMu.Unlock()
-	state, ok := r.txSetAcquire[txSetID]
-	if !ok {
-		return
-	}
-	state.attempts = 0
-	state.lastRequest = time.Time{}
-}
-
-// sweepStaleTxSetAcquireLocked drops entries older than txSetAcquireTTL.
-// Caller must hold r.txSetAcquireMu.
-func (r *Router) sweepStaleTxSetAcquireLocked() {
-	cutoff := time.Now().Add(-txSetAcquireTTL)
-	for id, state := range r.txSetAcquire {
-		if state.lastUpdate.Before(cutoff) {
-			delete(r.txSetAcquire, id)
-		}
-	}
-}
-
-// retryStalledTxSetAcquires re-requests the still-missing nodes of any
-// in-flight tx-set acquisition whose inbound responses have gone quiet, and
-// sweeps entries past their TTL. It fires every 250ms.
-//
-// The inbound retry (handleTxSetData) only advances on an arriving
-// TMLedgerData. When a peer falls silent mid-acquire — or every reply is
-// throttled — nothing re-requests the remaining nodes, so the acquisition
-// stalls until the 60s TTL sweep; under load that drops the node into
-// wrongLedger and the mixed network below quorum. This timer is the missing
-// driver. It reuses the same throttle/attempt-cap/peer-exclusion knobs as
-// the inbound path so the two never compound into a request storm: because
-// the inbound path keeps lastRequest fresh while it is making progress, the
-// MinInterval gate keeps this timer out of an actively progressing acquire
-// and only fires it once responses stop arriving.
-//
-// Runs on the Run() message-loop goroutine (same as handleTxSetData), so
-// reading state.txMap here never races the inbound path.
-func (r *Router) retryStalledTxSetAcquires() {
-	now := time.Now()
-	type txSetKick struct {
-		id       consensus.TxSetID
-		nodeIDs  [][]byte
-		excluded map[uint64]bool
-		attempts int
-		missing  int
-	}
-	type txSetDrop struct {
-		id       consensus.TxSetID
-		attempts int
-		missing  int
-	}
-	type txSetComplete struct {
-		id     consensus.TxSetID
-		txMap  *shamap.SHAMap
-		filled int
-	}
-	var kicks []txSetKick
-	var drops []txSetDrop
-	var completes []txSetComplete
-
-	r.txSetAcquireMu.Lock()
-	r.sweepStaleTxSetAcquireLocked()
-	knobs := r.txSetRetryKnobs
-	for id, state := range r.txSetAcquire {
-		// Only re-trigger once the inbound path has been quiet for a full
-		// cadence window. An actively progressing acquire keeps
-		// lastRequest fresh and is skipped here.
-		if !state.lastRequest.IsZero() && now.Sub(state.lastRequest) < knobs.MinInterval {
-			continue
-		}
-		missing := state.txMap.GetMissingNodes(256, nil)
-		if len(missing) == 0 {
-			// Tree is complete; the next inbound TMLedgerData (or a prior
-			// completion path) finalises it. Nothing to request.
-			continue
-		}
-		// Before re-requesting from peers, source any still-missing tx-leaf
-		// nodes from our own pending pool — the same local fill the inbound
-		// path performs (handleTxSetData). For tnTRANSACTION_NM the leaf hash
-		// equals the tx ID, so a single GetTx resolves the leaf. This avoids
-		// asking a peer for a tx we already hold; if the pool has grown since
-		// the last inbound reply it can complete the tree outright. GetTx is a
-		// local pool read and txMap is owned by this (Run) goroutine, so the
-		// fill is safe under txSetAcquireMu.
-		filled := 0
-		for _, m := range missing {
-			blob, getErr := r.adaptor.GetTx(consensus.TxID(m.Hash))
-			if getErr != nil || len(blob) == 0 {
-				continue
-			}
-			wire := txLeafWire(blob)
-			if addErr := state.txMap.AddKnownNode(m.Hash, wire); addErr == nil {
-				filled++
-			}
-		}
-		if filled > 0 {
-			_ = state.txMap.FinishSync()
-			missing = state.txMap.GetMissingNodes(256, nil)
-		}
-		if len(missing) == 0 {
-			// Completed from the local pool. Feed the engine just like the
-			// inbound completion path — peers are silent, so nothing else
-			// will. Finalised after the lock is released.
-			completes = append(completes, txSetComplete{id: id, txMap: state.txMap, filled: filled})
-			continue
-		}
-		if state.attempts >= knobs.MaxAttempts {
-			delete(r.txSetAcquire, id)
-			drops = append(drops, txSetDrop{id: id, attempts: state.attempts, missing: len(missing)})
-			continue
-		}
-		state.attempts++
-		state.lastRequest = now
-		excluded := buildExcludedPeers(state.peerNonProgress, knobs.PeerNonProgressThreshold)
-		nodeIDs := missingNodeIDs(missing)
-		kicks = append(kicks, txSetKick{id: id, nodeIDs: nodeIDs, excluded: excluded, attempts: state.attempts, missing: len(missing)})
-	}
-	r.txSetAcquireMu.Unlock()
-
-	for _, c := range completes {
-		r.finalizeLocalFilledTxSet(c.id, c.txMap, c.filled)
-	}
-	for _, d := range drops {
-		// Drop rather than mark failed: if consensus still needs the set,
-		// the next inbound TMLedgerData / MarkTxSetStillNeeded starts a
-		// fresh acquire (mirrors handleTxSetData's max-attempts handling).
-		r.logger.Info("tx-set sync: max attempts exceeded (timer)",
-			"t", "consensus", "event", "txset-reject",
-			"txset", fmt.Sprintf("%x", d.id[:8]),
-			"attempts", d.attempts,
-			"missing", d.missing,
-		)
-	}
-	for _, k := range kicks {
-		r.logger.Info("tx-set sync: timer re-trigger",
-			"t", "consensus", "event", "txset-timer-retry",
-			"txset", fmt.Sprintf("%x", k.id[:8]),
-			"missing", k.missing,
-			"attempts", k.attempts,
-			"excluded_peers", len(k.excluded),
-		)
-		if err := r.adaptor.RequestTxSetMissingNodes(k.id, k.nodeIDs, k.excluded); err != nil {
-			r.logger.Info("tx-set sync: timer missing-nodes request failed",
-				"t", "consensus", "event", "txset-reject",
-				"txset", fmt.Sprintf("%x", k.id[:8]),
-				"error", err.Error())
-		}
-	}
-}
-
-// finalizeLocalFilledTxSet feeds the engine a tx-set whose SHAMap the retry
-// timer completed from the local pending pool, mirroring the inbound completion
-// path in handleTxSetData. Runs on the Run() message-loop goroutine. The engine
-// re-checks the tx-set ID, so a stale or duplicate set is rejected with a log
-// rather than corrupting state.
-func (r *Router) finalizeLocalFilledTxSet(txSetID consensus.TxSetID, txMap *shamap.SHAMap, filled int) {
-	blobs := make([][]byte, 0)
-	if err := txMap.ForEach(func(item *shamap.Item) bool {
-		blobs = append(blobs, item.Data())
-		return true
-	}); err != nil {
-		r.deleteTxSetAcquire(txSetID)
-		return
-	}
-	r.deleteTxSetAcquire(txSetID)
-	if len(blobs) == 0 {
-		return
-	}
-	r.logger.Info("tx-set sync: completed via local pool (timer)",
-		"t", "consensus", "event", "txset-local-fill",
-		"txset", fmt.Sprintf("%x", txSetID[:8]),
-		"filled", filled,
-		"tx_count", len(blobs))
-	r.submitTxSetToEngine(txSetID, blobs)
-}
-
-// isShamapRootNodeID matches the SHAMap root wire encoding (33 zero bytes
-// = zero path + depth=0).
-func isShamapRootNodeID(b []byte) bool {
-	if len(b) != shamap.NodeIDSize {
-		return false
-	}
-	for _, by := range b {
-		if by != 0 {
-			return false
-		}
-	}
-	return true
-}
-
-func (r *Router) handleStatusChange(msg *peermanagement.InboundMessage) {
-	decoded, err := message.Decode(message.TypeStatusChange, msg.Payload)
-	if err != nil {
-		r.logger.Warn("failed to decode status_change", "error", err, "peer", msg.PeerID)
-		return
-	}
-	sc, ok := decoded.(*message.StatusChange)
-	if !ok {
-		return
-	}
-
-	r.logger.Info("peer status change",
-		"peer", msg.PeerID,
-		"status", sc.NewStatus,
-		"event", sc.NewEvent,
-		"ledger_seq", sc.LedgerSeq,
-		"needs_sync", r.adaptor.NeedsInitialSync(),
-	)
-
-	if sc.LedgerSeq > 0 {
-		var peerHash [32]byte
-		if len(sc.LedgerHash) == 32 {
-			copy(peerHash[:], sc.LedgerHash)
-		}
-		var parentHash [32]byte
-		if len(sc.LedgerHashPrevious) == 32 {
-			copy(parentHash[:], sc.LedgerHashPrevious)
-		}
-
-		r.peersMu.Lock()
-		r.peerStates[msg.PeerID] = &peerLedgerState{
-			LedgerSeq:  sc.LedgerSeq,
-			LedgerHash: peerHash,
-		}
-		r.peersMu.Unlock()
-
-		// Surface the peer's reported LCL to the adaptor so the
-		// engine's getNetworkLedger can consider it as a vote even
-		// when no proposal has (yet) arrived from this peer.
-		r.adaptor.UpdatePeerLCL(uint64(msg.PeerID), consensus.LedgerID(peerHash))
-
-		// During initial sync, fetch the full ledger from the peer.
-		// Don't adopt with synthetic headers — wait for real state data.
-		if r.adaptor.NeedsInitialSync() && sc.LedgerSeq > 1 {
-			r.startLedgerAcquisition(sc.LedgerSeq, peerHash, uint64(msg.PeerID))
-			return
-		}
-
-		// When in Full mode and significantly behind (gap > 2), acquire the
-		// latest ledger from the peer but stay in Full mode so we keep
-		// participating in consensus.
-		if r.adaptor.GetOperatingMode() == consensus.OpModeFull && sc.LedgerSeq > 1 {
-			svc := r.adaptor.LedgerService()
-			if svc != nil {
-				ourSeq := svc.GetClosedLedgerIndex()
-				if sc.LedgerSeq > ourSeq+2 {
-					r.logger.Warn("behind network while in Full mode, catching up",
-						"our_seq", ourSeq,
-						"peer_seq", sc.LedgerSeq,
-						"gap", sc.LedgerSeq-ourSeq,
-					)
-					r.startLedgerAcquisition(sc.LedgerSeq, peerHash, uint64(msg.PeerID))
-					return
-				}
-			}
-		}
-
-		// While not in Full mode, keep acquiring from peers until
-		// we're within 1 ledger of the network.
-		if r.adaptor.GetOperatingMode() != consensus.OpModeFull && sc.LedgerSeq > 1 {
-			svc := r.adaptor.LedgerService()
-			if svc != nil {
-				ourSeq := svc.GetClosedLedgerIndex()
-				if sc.LedgerSeq > ourSeq+1 {
-					r.startLedgerAcquisition(sc.LedgerSeq, peerHash, uint64(msg.PeerID))
-					return
-				}
-			}
-		}
-
-		// Hash-divergence catch-up. A late-join node (or a node whose
-		// consensus ran in isolation while disconnected) can end up at
-		// the same seq as its peers but with a different ledger hash.
-		// The seq-based branches above don't fire because ourSeq ==
-		// peerSeq; we need to detect that our LCL hash differs from the
-		// peer's and acquire theirs. Only fire if we're NOT already
-		// acquiring that hash (startLedgerAcquisition dedupes internally
-		// via the replayer / acquisition-registry guards, but checking
-		// here saves a lookup in the hot path).
-		svc := r.adaptor.LedgerService()
-		if svc != nil && sc.LedgerSeq > 1 && len(sc.LedgerHash) == 32 {
-			closed := svc.GetClosedLedger()
-			if closed != nil {
-				ourSeq := closed.Sequence()
-				ourHash := closed.Hash()
-				if ourSeq == sc.LedgerSeq && ourHash != peerHash {
-					r.logger.Warn("ledger hash divergence at same seq, acquiring peer's ledger",
-						"seq", sc.LedgerSeq,
-						"our_hash", fmt.Sprintf("%x", ourHash[:8]),
-						"peer_hash", fmt.Sprintf("%x", peerHash[:8]),
-						"peer", msg.PeerID,
-					)
-					r.startLedgerAcquisition(sc.LedgerSeq, peerHash, uint64(msg.PeerID))
-					return
-				}
-			}
-		}
-
-		r.checkBehind(sc.LedgerSeq, peerHash, uint64(msg.PeerID))
-	}
-}
-
-// startLedgerAcquisition picks the best available ledger-acquisition
-// strategy for the given target. When we have the parent ledger locally
-// and the peer advertises ledger-replay, the bandwidth-efficient
-// replay-delta protocol is preferred (one request returns header + every
-// tx blob); otherwise we fall back to the legacy mtGET_LEDGER
-// header+state walk.
-//
-// This is currently the only driver of startReplayDeltaAcquisition: it
-// handles a single target ledger per call. The Replayer coordinator
-// supports concurrent acquisitions across many hashes, but the policy
-// layer that walks a range (e.g., backward from a peer's tip via
-// ParentHash) is a follow-up item.
-func (r *Router) startLedgerAcquisition(seq uint32, hash [32]byte, peerID uint64) {
-	// Unified dedup across BOTH acquisition paths. A prior fix only
-	// checked r.replayer.Has(hash); that still allowed the cross-path
-	// race where two status changes at the same seq with different
-	// hashes armed both a replay-delta AND a legacy acquisition
-	// simultaneously, with adoption order then deciding which won. The
-	// single-point-of-truth check is a deliberate narrowing: a tighter
-	// guarantee that the same hash can't acquire through both paths.
-	if r.isAcquiring(hash) {
-		return
-	}
-
-	parent := r.adaptor.GetParentLedgerForReplay(seq)
-	if parent != nil && r.adaptor.PeerSupportsReplay(peerID) {
-		if err := r.startReplayDeltaAcquisition(seq, hash, peerID, parent); err == nil {
-			return
-		}
-		// Fall through to the legacy path on issue failure.
-	}
-	r.startLedgerAcquisitionLegacy(seq, hash, peerID)
-}
-
-// isAcquiring reports whether an acquisition — replay-delta or legacy
-// — is currently in flight for the given ledger hash. Used as the
-// single dedup entry point so a race between a replay-delta and a
-// legacy acquisition for the same hash is impossible.
-func (r *Router) isAcquiring(hash [32]byte) bool {
-	if r.replayer.Has(hash) {
-		return true
-	}
-	if r.fetchTracker.Find(hash) != nil {
-		return true
-	}
-	return false
-}
-
-// startReplayDeltaAcquisition registers a new acquisition with the
-// Replayer coordinator and issues the corresponding
-// mtREPLAY_DELTA_REQUEST.
-//
-// Returns ErrAcquisitionExists if a request for the same hash is
-// already in flight (caller should drop the duplicate), ErrCapacityFull
-// if the coordinator is at cap (caller falls back to legacy), or the
-// wire-send error if the request itself failed (coordinator slot is
-// freed before returning so the caller can retry).
-func (r *Router) startReplayDeltaAcquisition(seq uint32, hash [32]byte, peerID uint64, parent *ledger.Ledger) error {
-	rd, err := r.replayer.Acquire(hash, peerID, parent)
-	if err != nil {
-		return err
-	}
-	_ = rd // retained in replayer; HandleResponse retrieves it on reply.
-	r.logger.Info("starting replay delta acquisition",
-		"seq", seq,
-		"hash", fmt.Sprintf("%x", hash[:8]),
-		"peer", peerID,
-	)
-	if err := r.adaptor.RequestReplayDelta(peerID, hash); err != nil {
-		r.logger.Warn("failed to request replay delta from peer", "error", err)
-		r.replayer.Abandon(hash)
-		return err
-	}
-	return nil
-}
-
-// startLedgerAcquisitionLegacy requests the full ledger (header + state
-// tree) from a peer using the legacy mtGET_LEDGER protocol. This is the
-// fallback path when the parent isn't locally available or replay-delta
-// verification fails.
-//
-// Callers that enter via startLedgerAcquisition already consult
-// isAcquiring across both paths — but we still re-check here because
-// maintenanceTick and the replay-delta fallback paths can enter
-// directly, bypassing the unified entry point.
-func (r *Router) startLedgerAcquisitionLegacy(seq uint32, hash [32]byte, peerID uint64) {
-	// Safety net: if a replay-delta for the same hash is still
-	// registered, don't start a legacy on top of it — one path is
-	// always enough.
-	if r.replayer.Has(hash) {
-		return
-	}
-
-	_, created := r.fetchTracker.GetOrCreate(hash, func() *inbound.Ledger {
-		return inbound.New(hash, seq, peerID, r.logger)
-	})
-	if !created {
-		// Already acquiring this hash (consensus or a prior arm).
-		return
-	}
-
-	r.logger.Info("starting ledger acquisition (legacy)",
-		"seq", seq,
-		"hash", fmt.Sprintf("%x", hash[:8]),
-		"peer", peerID,
-	)
-
-	if err := r.adaptor.RequestLedgerBaseFromPeer(peerID, hash, seq); err != nil {
-		r.logger.Warn("failed to request ledger base from peer", "error", err)
-		r.fetchTracker.Remove(hash, false)
-	}
-}
-
-// FetchInfo returns the inbound-ledger acquisition snapshot served by the
-// fetch_info RPC. Safe to call from any goroutine.
-func (r *Router) FetchInfo() map[string]any {
-	return r.fetchTracker.Info()
-}
-
-// ClearFetchInfo resets the acquisition counters and recent-failure history,
-// backing fetch_info's `clear` param.
-func (r *Router) ClearFetchInfo() {
-	r.fetchTracker.Clear()
-}
-
-// RequestLedger triggers (or joins) a generic acquisition of a ledger from
-// peers, backing the ledger_request RPC. When hash is zero the target is
-// resolved from the validated ledger's skip list, and a ReasonGeneric
-// acquisition is started (or the in-flight one reused). started=true while
-// an acquisition is in flight; (nil,false,false) when the target can't be
-// resolved or no peer is available.
-//
-// reference distinguishes the two acquiring shapes: false when the
-// snapshot is the target ledger itself; true when it is a 256-aligned
-// reference ledger being fetched only to learn the target's hash.
-//
-// Safe to call from an RPC goroutine: the registry and each acquisition guard
-// their own state.
-func (r *Router) RequestLedger(hash [32]byte, seq uint32) (acquiring map[string]any, started, reference bool) {
-	// Don't acquire history online-delete has reclaimed: rippled's
-	// LedgerMaster::shouldAcquire refuses to fetch a missing ledger below
-	// minimumOnline. Re-fetching it would only feed the rotator another
-	// delete. Forward catch-up / validation acquisitions are above the
-	// validated tip (≥ floor) so they never hit this gate.
-	if seq != 0 && r.belowFloor(seq) {
-		r.logger.Debug("ledger_request declined: below online-delete floor",
-			"seq", seq, "floor", r.floor.MinimumOnline())
-		return nil, false, false
-	}
-	if hash == ([32]byte{}) {
-		if seq == 0 {
-			return nil, false, false
-		}
-		svc := r.adaptor.LedgerService()
-		if svc == nil {
-			return nil, false, false
-		}
-		vl := svc.GetValidatedLedger()
-		if vl == nil {
-			return nil, false, false
-		}
-		h, ok, err := vl.HashOfSeq(seq)
-		if err != nil {
-			return nil, false, false
-		}
-		if !ok {
-			// seq is past the rolling window and not 256-aligned, so its hash
-			// isn't directly in the validated ledger. Resolve it through a
-			// 256-aligned reference ledger whose hash IS enshrined in the skip
-			// list.
-			refIndex := getCandidateLedger(seq)
-			refHash, refOK, err := vl.HashOfSeq(refIndex)
-			if err != nil || !refOK {
-				return nil, false, false
-			}
-			refLedger, err := svc.GetLedgerByHash(refHash)
-			if err != nil || refLedger == nil {
-				// We lack the reference ledger needed to learn the target's
-				// hash — acquire it and report it as the in-flight reference.
-				if snap, ok := r.startGenericAcquisition(refHash, refIndex); ok {
-					return snap, true, true
-				}
-				return nil, false, false
-			}
-			h, ok, err = refLedger.HashOfSeq(seq)
-			if err != nil || !ok {
-				return nil, false, false
-			}
-		}
-		hash = h
-	}
-
-	if snap, ok := r.startGenericAcquisition(hash, seq); ok {
-		return snap, true, false
-	}
-	return nil, false, false
-}
-
-// startGenericAcquisition begins (or joins) a ReasonGeneric acquisition for
-// hash, issuing a base fetch from a selected peer only when it creates a fresh
-// one. Returns the acquisition snapshot, or ok=false when no peer is available
-// or the initial fetch could not be issued. The fetchTracker's GetOrCreate is
-// atomic, so a concurrent consensus catch-up arming the same hash is joined
-// rather than duplicated.
-func (r *Router) startGenericAcquisition(hash [32]byte, seq uint32) (map[string]any, bool) {
-	if il := r.fetchTracker.Find(hash); il != nil {
-		return inbound.AcquisitionJSON(il.Snapshot()), true
-	}
-
-	peerID, ok := r.selectAcquisitionPeer(seq)
-	if !ok {
-		return nil, false
-	}
-
-	il, created := r.fetchTracker.GetOrCreate(hash, func() *inbound.Ledger {
-		return inbound.NewGeneric(hash, seq, peerID, r.logger)
-	})
-	if created {
-		r.logger.Info("starting ledger acquisition (generic, ledger_request)",
-			"seq", seq,
-			"hash", fmt.Sprintf("%x", hash[:8]),
-			"peer", peerID,
-		)
-		if err := r.adaptor.RequestLedgerBaseFromPeer(peerID, hash, seq); err != nil {
-			r.logger.Warn("ledger_request: failed to request ledger base", "error", err)
-			r.fetchTracker.Remove(hash, false)
-			return nil, false
-		}
-	}
-	return inbound.AcquisitionJSON(il.Snapshot()), true
-}
-
-// getCandidateLedger rounds seq up to the next multiple of 256 — the nearest
-// ancestor whose hash is enshrined in the historical skip list and is therefore
-// easy to resolve, then close enough (within 256) to hold seq's hash in its own
-// rolling list.
-func getCandidateLedger(seq uint32) uint32 {
-	return (seq + 255) &^ 255
-}
-
-// selectAcquisitionPeer picks a connected peer to fetch a ledger from,
-// preferring one whose reported ledger is at or beyond the target sequence
-// (and therefore likely to hold it). When seq is unknown (0) or no peer is far
-// enough along, it falls back to any connected peer. Returns (0,false) when no
-// peer has reported a ledger state.
-func (r *Router) selectAcquisitionPeer(seq uint32) (uint64, bool) {
-	r.peersMu.RLock()
-	defer r.peersMu.RUnlock()
-
-	var fallback uint64
-	var haveFallback bool
-	for pid, st := range r.peerStates {
-		if !haveFallback {
-			fallback, haveFallback = uint64(pid), true
-		}
-		if seq == 0 || st.LedgerSeq >= seq {
-			return uint64(pid), true
-		}
-	}
-	return fallback, haveFallback
-}
-
-// handleReplayDeltaResponse verifies an inbound mtREPLAY_DELTA_RESPONSE
-// against its matching in-flight acquisition (routed by ledger hash)
-// and adopts the resulting ledger. On verification or apply failure the
-// acquisition is abandoned and the legacy path is started for the same
-// target. Unsolicited/stale responses (no matching acquisition) are
-// silently dropped — a normal race when a peer batch-forwards replies
-// after we've already moved on.
-func (r *Router) handleReplayDeltaResponse(msg *peermanagement.InboundMessage) {
-	decoded, err := message.Decode(message.TypeReplayDeltaResponse, msg.Payload)
-	if err != nil {
-		r.logger.Debug("failed to decode replay delta response", "error", err, "peer", msg.PeerID)
-		r.adaptor.IncPeerBadData(uint64(msg.PeerID), "replay-delta-resp-decode")
-		return
-	}
-	resp, ok := decoded.(*message.ReplayDeltaResponse)
-	if !ok || resp == nil {
-		return
-	}
-
-	// A delta acquired by an active LedgerReplayTask is owned by that
-	// task and never re-enters the generic InboundLedger flow.
-	if r.routeDeltaToActiveTask(resp) {
-		return
-	}
-
-	rd, err := r.replayer.HandleResponse(resp)
-	if errors.Is(err, inbound.ErrNoMatchingAcquisition) {
-		// Stale or unsolicited — drop silently without charging the
-		// peer. A misbehaving peer sending genuinely bogus data would
-		// fail its ACTIVE acquisition's verifier (branch below), which
-		// IS attributed via IncPeerBadData.
-		r.logger.Debug("replay delta response with no matching acquisition",
-			"peer", msg.PeerID)
-		return
-	}
-	if err != nil {
-		// Verification failed. rd is still registered in the Replayer so
-		// we can read its provenance before abandoning the slot.
-		seq := rd.Seq()
-		hash := rd.Hash()
-		peerID := rd.PeerID()
-		r.replayer.Abandon(hash)
-		r.logger.Warn("replay delta verification failed; falling back to legacy",
-			"seq", seq,
-			"hash", fmt.Sprintf("%x", hash[:8]),
-			"peer", peerID,
-			"error", err,
-		)
-		r.adaptor.IncPeerBadData(peerID, "replay-delta-verify")
-		r.startLedgerAcquisitionLegacy(seq, hash, peerID)
-		return
-	}
-
-	// GotResponse verified the header hash and the tx-map root. Apply
-	// re-derives the post-state by replaying every tx through the
-	// engine against a mutable copy of the parent's state, then
-	// verifies the resulting AccountHash matches the target header —
-	// the only proof we have that our engine produced the right state.
-	// Without this step the adopted ledger would carry the parent's
-	// stale state map, breaking consensus on the next round.
-	parent := rd.Parent()
-	engineCfg := r.adaptor.EngineConfigForReplay(parent)
-	derived, err := rd.Apply(engineCfg)
-	if err != nil {
-		seq := rd.Seq()
-		hash := rd.Hash()
-		peerID := rd.PeerID()
-		r.replayer.Abandon(hash)
-		// DO NOT charge the peer here. GotResponse already verified the
-		// peer's header hash and tx-map root; a subsequent Apply failure
-		// means OUR engine produced a divergent AccountHash — an engine
-		// bug, not peer misbehavior. Charging here would wrongly evict
-		// honest peers for our bugs.
-		r.logger.Error("ENGINE DIVERGENCE: replay delta apply failed; falling back to legacy",
-			"seq", seq,
-			"hash", fmt.Sprintf("%x", hash[:8]),
-			"peer", peerID,
-			"error", err,
-		)
-		r.startLedgerAcquisitionLegacy(seq, hash, peerID)
-		return
-	}
-	peerID := rd.PeerID()
-	r.replayer.Complete(rd.Hash())
-	if err := r.adoptVerifiedLedger(derived, peerID); err != nil {
-		r.logger.Warn("failed to adopt replay-delta ledger", "error", err)
-	}
-}
-
-// adoptVerifiedLedger commits a ledger reconstructed from a verified replay
-// delta, installing the peer-provided tx-blob tree alongside the state map.
-// Routes through SubmitHeldAdoption so out-of-order arrivals are stashed by
-// awaited parent seq; on stash we arm a backward-chain acquisition for the
-// parent.
-func (r *Router) adoptVerifiedLedger(l *ledger.Ledger, peerID uint64) error {
-	svc := r.adaptor.LedgerService()
-	if svc == nil {
-		return errors.New("no ledger service")
-	}
-	hdr := l.Header()
-	stateMap, err := l.StateMapSnapshot()
-	if err != nil {
-		return fmt.Errorf("snapshot state map: %w", err)
-	}
-	// Pass the verified tx map through so the adopted ledger carries
-	// real transactions — without this, tx/tx_history/account_tx RPCs
-	// can't answer for replay-delta-adopted ledgers and we can't
-	// re-serve the replay-delta to other peers.
-	txMap, err := l.TxMapSnapshot()
-	if err != nil {
-		return fmt.Errorf("snapshot tx map: %w", err)
-	}
-	// context.TODO: adoptVerifiedLedger is reached from a peer-message
-	// handler stack that does not currently carry a context. Threading
-	// one through the message-dispatch chain is tracked separately from
-	// this issue (#185).
-	res, err := svc.SubmitHeldAdoption(context.TODO(), &hdr, stateMap, txMap)
-	if err != nil {
-		return fmt.Errorf("adopt with state: %w", err)
-	}
-	if r.adaptor.GetOperatingMode() < consensus.OpModeTracking {
-		r.adaptor.SetOperatingMode(consensus.OpModeTracking)
-	}
-	r.logger.Info("adopted ledger via replay delta",
-		"seq", hdr.LedgerIndex,
-		"hash", fmt.Sprintf("%x", hdr.Hash[:8]),
-	)
-	// Notify the consensus engine so it can flip out of
-	// ModeWrongLedger via Engine.OnLedger. Without this, the engine
-	// remains stuck in wrongLedger indefinitely after a successful
-	// inbound acquisition.
-	if r.engine != nil {
-		if err := r.engine.OnLedger(consensus.LedgerID(hdr.Hash), nil); err != nil {
-			r.logger.Debug("engine rejected adopted ledger", "error", err, "seq", hdr.LedgerIndex)
-		}
-	}
-	if res.Stashed {
-		r.armParentAcquisition(svc, res.ParentSeq, res.ParentHash, peerID)
-	}
-	return nil
-}
-
-// maybeAcquireFromValidation arms inbound acquisition for a ledger attested
-// by a single TRUSTED validation, before the hash reaches quorum. It is the
-// non-quorum counterpart to armValidationStashAcquisition, acquiring the
-// ledger on EVERY trusted current validation when we don't already have it
-// — quorum is not required. With only the quorum-gated path, a node below
-// quorum (3 of 4 trusted validators on the network tip) never fetched that
-// tip and stalled in the wrongLedger chase loop.
-//
-// This only ACQUIRES. Advancing validatedLedger still flows through the
-// quorum gate (onFullyValidated → SetValidatedLedger), so a sub-quorum
-// fetch cannot move our validated tip and carries no state-divergence
-// risk; it just makes the ledger locally available so the node can rejoin
-// consensus on the network's chain instead of holding no position. The
-// deliberately-kept peer-LCL trusted-backing gate and the quorum gate are
-// untouched.
-func (r *Router) maybeAcquireFromValidation(v *consensus.Validation, originPeer uint64) {
-	if v == nil || v.LedgerSeq == 0 {
-		return
-	}
-	// Only trusted validators steer chain selection.
-	if !r.adaptor.IsTrusted(v.NodeID) {
-		return
-	}
-	svc := r.adaptor.LedgerService()
-	if svc == nil {
-		return
-	}
-	// Gate on the VALIDATED tip, never the closed/built tip — same rationale
-	// as armValidationStashAcquisition: a node that ran its closed chain
-	// ahead would otherwise skip the acquire and stay stuck on the wrong
-	// chain.
-	if v.LedgerSeq <= svc.GetValidatedLedgerIndex() {
-		return
-	}
-	hash := [32]byte(v.LedgerID)
-	// Already have it (built or adopted) — nothing to fetch.
-	if l, err := svc.GetLedgerByHash(hash); err == nil && l != nil {
-		return
-	}
-	// startLedgerAcquisition dedupes via isAcquiring, but checking here keeps
-	// the hot path quiet when many validations name the same hash.
-	if r.isAcquiring(hash) {
-		return
-	}
-	r.logger.Info("acquiring ledger from trusted validation",
-		"t", "consensus",
-		"event", "validation-acquire",
-		"seq", v.LedgerSeq,
-		"hash", fmt.Sprintf("%x", hash[:8]),
-		"peer", originPeer,
-	)
-	r.startLedgerAcquisition(v.LedgerSeq, hash, originPeer)
-}
-
-// armValidationStashAcquisition arms inbound acquisition for a (seq, hash)
-// that SetValidatedLedger stashed. Prefers a peer advertising LCL >= seq,
-// falls back to any tracked peer.
-func (r *Router) armValidationStashAcquisition(seq uint32, hash [32]byte) {
-	defer func() {
-		if rv := recover(); rv != nil {
-			r.logger.Error("armValidationStashAcquisition panic recovered",
-				"seq", seq,
-				"hash", fmt.Sprintf("%x", hash[:8]),
-				"panic", rv,
-			)
-		}
-	}()
-	if seq == 0 {
-		return
-	}
-	svc := r.adaptor.LedgerService()
-	if svc == nil {
-		return
-	}
-	// Skip only when seq is at or below the last *validated* ledger.
-	// Gating on the closed-ledger index instead silently swallowed
-	// recovery for a node that had run ahead on a private chain: when the
-	// validation tracker observed quorum on canonical seq=N with a
-	// different hash than our local seq=N, the acquire was skipped because
-	// closedSeq >> validatedSeq, leaving us stuck on the private chain
-	// forever.
-	if seq <= svc.GetValidatedLedgerIndex() {
-		return
-	}
-
-	// Walk peers in ID order so the chosen peer (and the emitted log)
-	// is reproducible across runs. Any peer with the hash can serve it.
-	r.peersMu.RLock()
-	peerIDs := make([]peermanagement.PeerID, 0, len(r.peerStates))
-	for pid := range r.peerStates {
-		peerIDs = append(peerIDs, pid)
-	}
-	slices.Sort(peerIDs)
-	var (
-		preferredPeerID uint64
-		fallbackPeerID  uint64
-	)
-	for _, pid := range peerIDs {
-		st := r.peerStates[pid]
-		if fallbackPeerID == 0 {
-			fallbackPeerID = uint64(pid)
-		}
-		if st != nil && st.LedgerSeq >= seq {
-			preferredPeerID = uint64(pid)
-			break
-		}
-	}
-	r.peersMu.RUnlock()
-	if preferredPeerID == 0 {
-		preferredPeerID = fallbackPeerID
-	}
-	if preferredPeerID == 0 {
-		return
-	}
-
-	r.logger.Info("arming acquisition for stashed validation",
-		"seq", seq,
-		"hash", fmt.Sprintf("%x", hash[:8]),
-		"preferred_peer", preferredPeerID,
-	)
-	r.startLedgerAcquisition(seq, hash, preferredPeerID)
-}
-
-// armParentAcquisition fires a backward-chain acquisition for the parent of
-// a stashed held-adoption candidate. Skips at-or-below closed (already
-// adopted or fork-dropped).
-func (r *Router) armParentAcquisition(svc *service.Service, parentSeq uint32, parentHash [32]byte, preferredPeerID uint64) {
-	if parentSeq == 0 {
-		return
-	}
-	if parentSeq <= svc.GetClosedLedgerIndex() {
-		return
-	}
-	r.logger.Info("arming backward-chain acquisition for stashed held-adoption parent",
-		"parent_seq", parentSeq,
-		"parent_hash", fmt.Sprintf("%x", parentHash[:8]),
-		"preferred_peer", preferredPeerID,
-	)
-	r.startLedgerAcquisition(parentSeq, parentHash, preferredPeerID)
-}
-
-// checkBehind decides what to do based on how far behind a peer
-// reports. Two outcomes:
-//
-//   - peerSeq <= ourSeq+1: we're caught up. If still in Tracking and
-//     our LCL hash matches peers' majority, transition to Full.
-//     Otherwise stay in Tracking — the hash-mismatch branch in
-//     handleStatusChange will have already fired the right acquisition.
-//   - peerSeq > ourSeq+1: we're behind by more than one ledger. Arm a
-//     single acquisition for the peer's tip. Subsequent status changes
-//     from peers will chain more acquisitions forward as we adopt each
-//     ledger and ourSeq advances.
-//
-// Only one acquisition fires per call. A faster "range walk" that
-// issues concurrent requests for every seq between ourLCL+1 and
-// peerSeq would need the intermediate ledger hashes, which we don't
-// know until each acquired header reveals its ParentHash; we rely on
-// forward status gossip instead. Replayer already supports concurrent
-// in-flight acquisitions, so switching to backward-walk later is a
-// localized change in this function.
-func (r *Router) checkBehind(peerSeq uint32, peerHash [32]byte, peerID uint64) {
-	svc := r.adaptor.LedgerService()
-	if svc == nil {
-		return
-	}
-
-	ourSeq := svc.GetClosedLedgerIndex()
-
-	// If we're caught up (gap ≤ 1) and not yet Full, transition to Full
-	// only if our LCL hash matches what the majority of peers report.
-	if peerSeq <= ourSeq+1 {
-		if r.adaptor.GetOperatingMode() == consensus.OpModeTracking {
-			if r.ourLCLMatchesPeers() {
-				r.logger.Info("caught up with network, transitioning to Full",
-					"our_seq", ourSeq,
-					"peer_seq", peerSeq,
-				)
-				r.adaptor.SetOperatingMode(consensus.OpModeFull)
-			} else {
-				r.logger.Info("caught up but LCL hash differs, staying in Tracking",
-					"our_seq", ourSeq,
-					"peer_seq", peerSeq,
-				)
-			}
-		}
-		return
-	}
-
-	r.logger.Info("behind network, acquiring peer tip",
-		"our_seq", ourSeq,
-		"peer_seq", peerSeq,
-		"gap", peerSeq-ourSeq,
-		"peer", peerID,
-	)
-
-	// Arm a real acquisition instead of broadcasting a bare
-	// mtGET_LEDGER. RequestLedgerByHashAndSeq would broadcast the
-	// request but never arm the InboundLedger state machine, so any
-	// response would arrive with no active consumer and be dropped.
-	// startLedgerAcquisition picks replay-delta or legacy per the
-	// routing policy and both paths install their own state machines.
-	r.startLedgerAcquisition(peerSeq, peerHash, peerID)
-}
-
-// ourLCLMatchesPeers checks if our closed ledger hash matches what the
-// majority of tracked peers report. Returns true if we have no peer data
-// (to avoid blocking startup).
-func (r *Router) ourLCLMatchesPeers() bool {
-	svc := r.adaptor.LedgerService()
-	if svc == nil {
-		return true
-	}
-	closedLedger := svc.GetClosedLedger()
-	if closedLedger == nil {
-		return true
-	}
-	ourHash := closedLedger.Hash()
-	ourSeq := svc.GetClosedLedgerIndex()
-
-	r.peersMu.RLock()
-	defer r.peersMu.RUnlock()
-
-	if len(r.peerStates) == 0 {
-		return true
-	}
-
-	matching := 0
-	total := 0
-	for _, ps := range r.peerStates {
-		if ps.LedgerSeq == ourSeq {
-			total++
-			if ps.LedgerHash == ourHash {
-				matching++
-			}
-		}
-	}
-
-	// If no peers at our seq, allow transition (they may have advanced)
-	if total == 0 {
-		return true
-	}
-
-	return matching > total/2
-}
-
-func (r *Router) handleLedgerData(msg *peermanagement.InboundMessage) {
-	decoded, err := message.Decode(message.TypeLedgerData, msg.Payload)
-	if err != nil {
-		r.logger.Warn("failed to decode ledger_data", "error", err, "peer", msg.PeerID)
-		r.adaptor.IncPeerBadData(uint64(msg.PeerID), "ledger-data-decode")
-		return
-	}
-	ld, ok := decoded.(*message.LedgerData)
-	if !ok {
-		return
-	}
-
-	var il *inbound.Ledger
-	if len(ld.LedgerHash) == 32 {
-		var h [32]byte
-		copy(h[:], ld.LedgerHash)
-		il = r.fetchTracker.Find(h)
-	}
-
-	r.logger.Info("received ledger data",
-		"peer", msg.PeerID,
-		"seq", ld.LedgerSeq,
-		"nodes", len(ld.Nodes),
-		"itype", ld.InfoType,
-		"has_inbound", il != nil,
-	)
-
-	// liTS_CANDIDATE response — feeds the engine via the tx-set path
-	// (consensus-time only).
-	if ld.InfoType == message.LedgerInfoTsCandidate {
-		r.handleTxSetData(ld, uint64(msg.PeerID))
-		return
-	}
-
-	if il != nil {
-		if r.handleInboundLedgerData(il, ld) {
-			return
-		}
-		// If handleInboundLedgerData returned false (e.g. GotBase failed),
-		// fall through to the legacy header-only adoption path
-	}
-
-	// During initial sync, try to adopt the ledger header from peers
-	if ld.InfoType == message.LedgerInfoBase && len(ld.Nodes) > 0 && r.adaptor.NeedsInitialSync() {
-		headerData := ld.Nodes[0].NodeData
-		if err := r.adaptor.AdoptLedgerFromHeader(headerData); err != nil {
-			r.logger.Debug("failed to adopt ledger header", "error", err, "peer", msg.PeerID)
-		} else {
-			r.logger.Info("adopted ledger from peer",
-				"seq", ld.LedgerSeq,
-				"peer", msg.PeerID,
-			)
-			return
-		}
-	}
-
-	if len(ld.LedgerHash) == 32 {
-		var ledgerID consensus.LedgerID
-		copy(ledgerID[:], ld.LedgerHash)
-
-		var payload []byte
-		for _, node := range ld.Nodes {
-			payload = append(payload, node.NodeData...)
-		}
-
-		if err := r.engine.OnLedger(ledgerID, payload); err != nil {
-			r.logger.Debug("engine rejected ledger data", "error", err, "peer", msg.PeerID)
-		}
-	}
-}
-
-// handleInboundLedgerData feeds LedgerData to the given InboundLedger
-// acquisition (already matched by hash in handleLedgerData). Returns true if
-// the data was consumed by the acquisition.
-func (r *Router) handleInboundLedgerData(il *inbound.Ledger, ld *message.LedgerData) bool {
-	if il == nil {
-		return false
-	}
-
-	switch ld.InfoType {
-	case message.LedgerInfoBase:
-		if len(ld.Nodes) < 2 {
-			// Response doesn't include root nodes — can't do full acquisition.
-			// Drop the acquisition and fall through to legacy adoption.
-			r.logger.Debug("inbound ledger: response has < 2 nodes, falling back", "nodes", len(ld.Nodes))
-			r.fetchTracker.Remove(il.Hash(), false)
-			return false
-		}
-		if err := il.GotBase(ld.Nodes); err != nil {
-			r.logger.Warn("inbound ledger: GotBase failed, falling back", "error", err)
-			r.adaptor.IncPeerBadData(il.PeerID(), "ledger-data-base")
-			r.fetchTracker.Remove(il.Hash(), false)
-			return false
-		}
-
-		if il.IsComplete() {
-			r.completeInboundLedger(il)
-			return true
-		}
-
-		// Request the missing state and transaction nodes in parallel.
-		r.requestMissingAcquisitionNodes(il)
-		return true
-
-	case message.LedgerInfoAsNode:
-		if err := il.GotStateNodes(ld.Nodes); err != nil {
-			r.logger.Warn("inbound ledger: GotStateNodes failed", "error", err)
-			r.adaptor.IncPeerBadData(il.PeerID(), "ledger-data-state")
-			return true
-		}
-
-		if il.IsComplete() {
-			r.completeInboundLedger(il)
-			return true
-		}
-
-		r.requestMissingAcquisitionNodes(il)
-		return true
-
-	case message.LedgerInfoTxNode:
-		if err := il.GotTransactionNodes(ld.Nodes); err != nil {
-			r.logger.Warn("inbound ledger: GotTransactionNodes failed", "error", err)
-			r.adaptor.IncPeerBadData(il.PeerID(), "ledger-data-tx")
-			return true
-		}
-
-		if il.IsComplete() {
-			r.completeInboundLedger(il)
-			return true
-		}
-
-		r.requestMissingAcquisitionNodes(il)
-		return true
-	}
-
-	return false
-}
-
-// requestMissingAcquisitionNodes asks the source peer for the outstanding
-// account-state and transaction tree nodes of the active acquisition,
-// requesting both trees in parallel. Each call is a no-op for a tree
-// already complete.
-func (r *Router) requestMissingAcquisitionNodes(il *inbound.Ledger) {
-	if nodeIDs := il.NeedsMissingNodeIDs(); len(nodeIDs) > 0 {
-		if err := r.adaptor.RequestStateNodes(il.PeerID(), il.Hash(), nodeIDs); err != nil {
-			r.logger.Warn("inbound ledger: failed to request state nodes", "error", err)
-		}
-	}
-	if nodeIDs := il.NeedsMissingTxNodeIDs(); len(nodeIDs) > 0 {
-		if err := r.adaptor.RequestTransactionNodes(il.PeerID(), il.Hash(), nodeIDs); err != nil {
-			r.logger.Warn("inbound ledger: failed to request tx nodes", "error", err)
-		}
-	}
-}
-
-// completeInboundLedger finalizes an InboundLedger acquisition and adopts the
-// ledger. A ReasonGeneric acquisition (RPC-driven, ledger_request) is persisted
-// for querying but does not flip operating mode or notify consensus, so an
-// arbitrary historical fetch can't disturb the active chain.
-func (r *Router) completeInboundLedger(il *inbound.Ledger) {
-	r.fetchTracker.Remove(il.Hash(), true)
-
-	h, stateMap, txMap, err := il.Result()
-	if err != nil {
-		r.logger.Warn("inbound ledger: failed to get result", "error", err)
-		return
-	}
-	peerID := il.PeerID()
-
-	svc := r.adaptor.LedgerService()
-	if svc == nil {
-		return
-	}
-
-	// The acquisition fetches the header, state map, and transaction map; txMap
-	// is nil only when the ledger has no transactions (empty tx tree), in which
-	// case the service installs the genesis-shaped empty tx map.
-	//
-	// Route through SubmitHeldAdoption so out-of-order catchup arrivals
-	// either fast-path (parent already present) or stash for cascade when the
-	// awaited parent lands. Legacy mtGET_LEDGER is sequential at the wire level
-	// today, but nothing in the protocol forbids interleaving — the held-queue
-	// is the correct seam regardless.
-	// context.TODO: same as adoptVerifiedLedger — reached from a peer-message
-	// handler stack with no plumbed context. See note there.
-	res, err := svc.SubmitHeldAdoption(context.TODO(), h, stateMap, txMap)
-	if err != nil {
-		r.logger.Warn("inbound ledger: failed to adopt with state", "error", err)
-		return
-	}
-
-	// A generic (RPC-driven) acquisition is now persisted and queryable; it
-	// must not advance operating mode or feed consensus.
-	if il.Reason() == inbound.ReasonGeneric {
-		r.logger.Info("acquired ledger (generic) with full state from peer",
-			"seq", h.LedgerIndex,
-			"hash", fmt.Sprintf("%x", h.Hash[:8]),
-		)
-		if res.Stashed {
-			r.armParentAcquisition(svc, res.ParentSeq, res.ParentHash, peerID)
-		}
-		return
-	}
-
-	// Only upgrade to Tracking if still in a lower mode.
-	// Never demote from Full — that would break consensus participation.
-	if r.adaptor.GetOperatingMode() < consensus.OpModeTracking {
-		r.adaptor.SetOperatingMode(consensus.OpModeTracking)
-	}
-	r.logger.Info("adopted ledger with full state from peer",
-		"seq", h.LedgerIndex,
-		"hash", fmt.Sprintf("%x", h.Hash[:8]),
-		"account_hash", fmt.Sprintf("%x", h.AccountHash[:8]),
-	)
-	// Notify the consensus engine so it can flip out of
-	// ModeWrongLedger via Engine.OnLedger, mirroring the replay-delta
-	// path in adoptVerifiedLedger.
-	if r.engine != nil {
-		if err := r.engine.OnLedger(consensus.LedgerID(h.Hash), nil); err != nil {
-			r.logger.Debug("engine rejected adopted ledger", "error", err, "seq", h.LedgerIndex)
-		}
-	}
-	if res.Stashed {
-		r.armParentAcquisition(svc, res.ParentSeq, res.ParentHash, peerID)
-	}
 }
